+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/server/files/FileStoreMonitor.java
@@ -21,6 +21,8 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.FileStore;
 import java.nio.file.Files;
+import java.nio.file.attribute.FileAttributeView;
+import java.nio.file.attribute.FileStoreAttributeView;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.concurrent.Executor;
@@ -74,7 +76,7 @@ public class FileStoreMonitor extends ActiveMQScheduledComponent {
                addStore(Files.getFileStore(file.toPath()));
             } catch (IOException e) {
                logger.warn("Error getting file store for " + file.getAbsolutePath() + ". Trying dummy FileStore", e);
-               addStore(new DummyFileStore());
+               addStore(new UnknownFileStore());
             }
          }
          return this;
@@ -155,6 +157,91 @@ public class FileStoreMonitor extends ActiveMQScheduledComponent {
       void over(FileStore store, double usage);
 
       void under(FileStore store, double usage);
+   }
+
+
+/**
+ * Dummy implementation of FileStore object. Its a workaround to handle the exception thrown from org.apache.activemq.artemis.core.server.files.FileStoreMonitor.addStore(File)
+ * when running on a RAMDISK machine or similar file system
+ *
+ */
+   class UnknownFileStore extends FileStore {
+
+      @Override
+      public String name() {
+         return "FS_UNKNOWN_NAME";
+      }
+
+      @Override
+      public String type() {
+         return "FS_UNKNOWN_TYPE";
+      }
+
+      @Override
+      public boolean isReadOnly() {
+         return false;
+      }
+
+      @Override
+      public long getTotalSpace() throws IOException {
+         return Long.MAX_VALUE;
+      }
+
+      @Override
+      public long getUsableSpace() throws IOException {
+         return Long.MAX_VALUE;
+      }
+
+      @Override
+      public long getUnallocatedSpace() throws IOException {
+         return Long.MAX_VALUE;
+      }
+
+      @Override
+      public String toString() {
+         return name();
+      }
+
+      /*
+      * (non-Javadoc)
+      *
+      * @see java.nio.file.FileStore#supportsFileAttributeView(java.lang.Class)
+      */
+      @Override
+      public boolean supportsFileAttributeView(Class<? extends FileAttributeView> type) {
+         return false;
+      }
+
+      /*
+      * (non-Javadoc)
+      *
+      * @see java.nio.file.FileStore#supportsFileAttributeView(java.lang.String)
+      */
+      @Override
+      public boolean supportsFileAttributeView(String name) {
+         return false;
+      }
+
+      /*
+      * (non-Javadoc)
+      *
+      * @see java.nio.file.FileStore#getFileStoreAttributeView(java.lang.Class)
+      */
+      @Override
+      public <V extends FileStoreAttributeView> V getFileStoreAttributeView(Class<V> type) {
+         return null;
+      }
+
+      /*
+      * (non-Javadoc)
+      *
+      * @see java.nio.file.FileStore#getAttribute(java.lang.String)
+      */
+      @Override
+      public Object getAttribute(String attribute) throws IOException {
+         return null;
+      }
+
    }
 
 }
